+++ src/pages/PostsManagerPage.tsx
@@ -2,15 +2,13 @@ import {
   Edit2,
   MessageSquare,
   Plus,
-  Search,
   ThumbsDown,
   ThumbsUp,
   Trash2,
 } from "lucide-react"
 import { useState } from "react"
-import { SortOrder, usePostQueryParams, usePostsQuery } from "../entities/post"
+import { usePostQueryParams, usePostsQuery } from "../entities/post"
 import { usePostsQueryProps } from "../entities/post/api/usePostsQuery"
-import { usePostTagsQuery } from "../entities/post/api/usePostTagsQuery"
 import { Author, Post } from "../entities/post/model/types"
 import { userApi } from "../entities/user/api/userApi"
 import { UserDTO } from "../entities/user/model/types"
@@ -21,12 +19,6 @@ import {
   CardContent,
   CardHeader,
   CardTitle,
-  Input,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
   Table,
   TableBody,
   TableCell,
@@ -39,6 +31,7 @@ import {
   PostAddDialog,
   PostDetailDialog,
   PostEditDialog,
+  PostFilterBar,
 } from "../widgets/post"
 import { Pagination } from "../widgets/ui/Pagination"
 import { UserModal } from "../widgets/user"
@@ -48,11 +41,10 @@ const PostsManager = () => {
     queryParams: { limit, search, skip, sortBy, sortOrder, tag: selectedTag },
     updateQueryParam,
   } = usePostQueryParams()
+  const [searchQuery, setSearchQuery] = useState(search)
 
   // 상태 관리
   const [selectedPost, setSelectedPost] = useState<Post | null>(null)
-
-  const [searchQuery, setSearchQuery] = useState(search)
 
   const [showAddDialog, setShowAddDialog] = useState(false)
   const [showEditDialog, setShowEditDialog] = useState(false)
@@ -62,13 +54,13 @@ const PostsManager = () => {
   const [showUserModal, setShowUserModal] = useState(false)
   const [selectedUser, setSelectedUser] = useState<UserDTO | null>(null)
 
-  const { data: tags = [] } = usePostTagsQuery()
-
   const payload: usePostsQueryProps = {
     limit,
     skip,
-    searchQuery,
+    search: searchQuery,
     selectedTag,
+    sortBy,
+    sortOrder,
   }
 
   const {
@@ -203,69 +195,14 @@ const PostsManager = () => {
       <CardContent>
         <div className="flex flex-col gap-4">
           {/* 검색 및 필터 컨트롤 */}
-          <div className="flex gap-4">
-            <div className="flex-1">
-              <div className="relative">
-                <Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
-                <Input
-                  placeholder="게시물 검색..."
-                  className="pl-8"
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  onKeyPress={(e) =>
-                    e.key === "Enter" &&
-                    updateQueryParam({ search: searchQuery })
-                  }
-                />
-              </div>
-            </div>
-            <Select
-              value={selectedTag}
-              onValueChange={(value) => {
-                updateQueryParam({ tag: value })
-              }}
-            >
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="태그 선택" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="all">모든 태그</SelectItem>
-                {tags.map((tag) => (
-                  <SelectItem key={tag.url} value={tag.slug}>
-                    {tag.slug}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-            <Select
-              value={sortBy}
-              onValueChange={(value) => updateQueryParam({ sortBy: value })}
-            >
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="정렬 기준" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="none">없음</SelectItem>
-                <SelectItem value="id">ID</SelectItem>
-                <SelectItem value="title">제목</SelectItem>
-                <SelectItem value="reactions">반응</SelectItem>
-              </SelectContent>
-            </Select>
-            <Select
-              value={sortOrder}
-              onValueChange={(value) =>
-                updateQueryParam({ sortOrder: value as SortOrder })
-              }
-            >
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="정렬 순서" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="asc">오름차순</SelectItem>
-                <SelectItem value="desc">내림차순</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
+          <PostFilterBar
+            searchQuery={searchQuery}
+            setSearchQuery={setSearchQuery}
+            selectedTag={selectedTag}
+            sortBy={sortBy}
+            sortOrder={sortOrder}
+            updateQueryParam={updateQueryParam}
+          />
 
           {/* 게시물 테이블 */}
           {isLoading ? (
@@ -298,7 +235,7 @@ const PostsManager = () => {
       <PostDetailDialog
         open={showPostDetailDialog}
         onOpenChange={setShowPostDetailDialog}
-        searchQuery={searchQuery}
+        search={searchQuery}
         selectedPost={selectedPost}
       />
