--- src/shared/model/paginationType.ts
@@ -1,4 +1,0 @@
-export type PaginationType = {
-  skip: number
-  limit: number
-}
