+++ src/shared/model/types.ts
@@ -1,0 +1,11 @@
+export type PaginationType = {
+  skip: number
+  limit: number
+}
+
+export type SortOrder = "asc" | "desc"
+
+export type SortType = {
+  sortBy: string
+  sortOrder: SortOrder
+}
