+++ src/widgets/post/index.ts
@@ -1,3 +1,4 @@
 export * from "./ui/PostAddDialog"
 export * from "./ui/PostDetailDialog"
 export * from "./ui/PostEditDialog"
+export * from "./ui/PostFilterBar"
